added separate directories and modified run script

+++ run.sh
@@ -1,10 +1,15 @@
  
 #! /bin/bash
 
-executable=$1
+directory=$1
+executable=`echo $directory | sed 's/\(.*\)\//\1/'`_main.out
 
 mkdir -p outputs
-rm -rf *.o
-gcc -c *.c
-gcc -o outputs/$1 *.o && outputs/$1
+rm -rf $directory/*.o
+gcc -c $directory/*.c
+gcc -o outputs/$executable *.o && outputs/$executable
+if [[ $? -ne 0 ]]; then
+    rm -rf *.o
+    exit 1
+fi
 rm -rf *.o
